Implement MmapSerializationSink and add benchmarks.
