+++ measureme/Cargo.toml
@@ -12,4 +12,4 @@ travis-ci = { repository = "rust-lang/measureme" }
 [dependencies]
 byteorder = "1.3"
 rustc-hash = "1.0.1"
-
+memmap = "0.7.0"
